Pushed v1.02 code
- The paulaSetLength() function now takes length in words, not bytes.
  This doesn't really change anything, but it's more correct.
- Fixed a very minor bug with VOL up/down button in Edit Op.
- Fixed a bug where the scopes could show the wrong volume for the
  tunetone waveform (sampler screen).
- Some code cleanup

+++ src/pt2_audio.h
@@ -29,7 +29,7 @@ void paulaStopDMA(uint8_t ch);
 void paulaStartDMA(uint8_t ch);
 void paulaSetPeriod(uint8_t ch, uint16_t period);
 void paulaSetVolume(uint8_t ch, uint16_t vol);
-void paulaSetLength(uint8_t ch, uint32_t len);
+void paulaSetLength(uint8_t ch, uint16_t len);
 void paulaSetData(uint8_t ch, const int8_t *src);
 void lockAudio(void);
 void unlockAudio(void);
